+++ src/routes/about.tsx
@@ -1,0 +1,20 @@
+/**
+ * About page for Orbite. (/about)
+ *
+ * @author Richard Nguyen <richard@richardhnguyen.com>
+ */
+
+import * as React from "react";
+
+import { CFC } from "@typings/react";
+import Layout from "@components/Layout";
+
+const About: CFC<HTMLDivElement> = () => {
+  return (
+    <Layout>
+      <h1>About Orbite</h1>
+    </Layout>
+  );
+};
+
+export default About;