+++ src/routes/index.ts
@@ -1,7 +1,11 @@
+import About from "./about";
 import Root from "./root";
+import Setting from "./setting";
 
 const exportedRoutes = {
   Root: Root,
+  About: About,
+  Setting: Setting,
 };
 
 export default exportedRoutes;