+++ src/routes/setting.tsx
@@ -1,0 +1,20 @@
+/**
+ * Setting page for Orbite. (/setting)
+ *
+ * @author Richard Nguyen <richard@richardhnguyen.com>
+ */
+
+import * as React from "react";
+
+import { CFC } from "@typings/react";
+import Layout from "@components/Layout";
+
+const Setting: CFC<HTMLDivElement> = () => {
+  return (
+    <Layout>
+      <h1>Setting</h1>
+    </Layout>
+  );
+};
+
+export default Setting;